Refactor MySQL table for users: Modify 'email' and 'name' columns
This commit refactors the MySQL table for users by modifying the 'email' and 'name' columns. The 'email' column is changed to have a maximum length of 255 characters and is set to be not null. The 'name' column is made optional by removing the not null constraint. These changes improve the flexibility and usability of the table for storing user information.